[SPARK-41681][CONNECT] Factor GroupedData out to group.py
### What changes were proposed in this pull request?

This PR factors `GroupedData` to group.py to be consistent with `pyspark.sql.group`.

### Why are the changes needed?

For consistency and readability.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Existing test cases should cover this.

Closes #39172 from HyukjinKwon/SPARK-41681.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
